feat: split out code, convert to typescript (#111)
* Splits implementations out from interfaces so we can depend on interfaces module without pulling in crypto, etc (fixes #110).  New implementation modules are:
  * libp2p-pubsub
  * libp2p-connection
  * libp2p-topology
* Converts code to typescript (fixes #101)
* Adds types for PeerData, PeerStore, Registrar and some other libp2p internals - these will need to be fleshed out properly in a subsequent PR
* No CJS, only ESM, only forwards
* Upgrades deps that have also gone ESM-only
* Renames package folders for consistency

I've tried to break some of the circular dependencies in the interfaces (e.g. libp2p depends on pubsub, which depends on libp2p, which depends on pubsub, etc).  This hasn't been possible in the tests so I've added `@ts-expect-error` where we access these properties with the aim to revisit it once this has been rolled up to libp2p itself.

BREAKING CHANGE: not all fields from concrete classes have been added to the interfaces, some adjustment may be necessary as this gets rolled out