fix: update dialer tests
The `mockUpgrader` is supposed to accept a `MultiaddrConnection` and
return a `Connection`, but it just passes through the `MultiaddrConnection`.

This is a problem as the two types are not compatible.  This shows up in our
code as our tests treat `Connection`s as if they are something you can pass
to `it-pipe` but they are not.

This PR fixes up the `mockUpgrader` to return something compatible with the
`Connection` interface.  It also removes a few of the dialer tests that were
only testing the newly introduced mock code.