feat: add observed address methods
When we observe an address that might be externally dialable (via
identify or uPNP NAT holepunching, etc), we need some way of signalling
that we've either verified it's externally dialable and that we are
safe to advertise it, or that it's not dialable and that we should
forget about it.

These methods are necessary to implement autonat.