fix!: remove peer id argument from deny dial multiaddr
Since we can now dial multiaddrs without a peer id attached, the peer
id may not be present on the multiaddr so we cannot have a required
peer id argument passed to `denyDialMultiaddr`.

If present it can be accessed by calling `getPeerId` on the multiaddr.

BREAKING CHANGE: the peer id argument has been removed from the `denyDialMultiaddr` method of the connection gater