feat: add peer:identify event to libp2p
This event is useful to evesdrop on network behaviour and helps to
debug situations where a remote peer is sending bad data in it's
identify response (for exapmle not including any routable addresses).

An event is better than opening a protocol stream and sending more
requests as remotes will frequently reset streams if there are too
many concurrent identify requests, etc.